Match type for pure_check_for_work
Before the reactor main loop a local pure_check_for_work std::function
is created on the stack. This is only ever passed to
the _idle_cpu_handler, which expects a
ss::noncopyable_function. The type mismatch means we need to clone the
std::function every time, creating a temporary
object to bind to the seastar function.

This causes the construction and destruction methods
to show up as a non-trivial portion of performance
profiles whenever the idle check occurs frequently enough.

Instead, just change the type to non-copyable function so the
reference can be passed directly without any copy.