chunked_fifo: add default ctor for basic_iterator
An empty `context` is created in the seastar wrapper for collectd
```
    auto ctxt = make_lw_shared<context>();
```
and the `context` takes an iterator of `metric_metadata_vector`.

After replacing the metadata vector with a chunked_fifo, the call to
create a `lw_shared_ptr<context>` will fail compilation because
`chunked_fifo::basic_iterator` does not have a default constuctor.

Therefore, this patch adds a default constructor to
`chunked_fifo::basic_iterator`.

Signed-off-by: NyaliaLui <nyalia@redpanda.com>