Fix edge case in memory sampler at OOM
In one place in the memory sampling code we assume a non-null pointer
but if the system is out of memory this pointer may be null.

When using abort-on-bad-alloc this scenario is extremely unlikely since
the first failed allocation would have to meet very specific conditions
for this to occur: be a small allocation and be sampled, which is
unlikely since larger allocations usually fail first (and small pools
ask the underlying allocator for at most 128K at a time).

Discovered while analyzing redpanda-data/redpanda#15711.